split address forms into their own components
